+++ examples/constpat/lib.c
@@ -8,19 +8,19 @@ extern void *data;
     snprintf(buf, sizeof buf, "wider_still(%d, %d, %d)", a, b, c);
     return buf;
 }*/
-char *narrow(int a)
+char *narrow(long int a)
 {
  	static char buf[4096];
     snprintf(buf, sizeof buf, "narrow(%d)", a); 
     return buf;  
 }
-char *narrowish(int a, int b)
+char *narrowish(long int a, long int b)
 {
  	static char buf[4096];
     snprintf(buf, sizeof buf, "narrowish(%d, %d)", a, b);
     return buf;
 }
-char *not_always(int a)
+char *not_always(long int a)
 {
  	static char buf[4096];
     snprintf(buf, sizeof buf, "not_always(%d)", a);
